+++ docs/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed typing error in racetrack CLI.
+  ([#455](https://github.com/TheRacetrack/racetrack/issues/455))
 
 ## [2.29.0] - 2024-04-25
 ### Added